feat(status): Implement `wl status`
Since `wl` uses `nmcli`, a new module is created for individual `nmcli`
commands.

`wl status` (or `wl s` with alias) is implemented by printing the wifi
status and connected networks to stdout.

Loopback interface is not stripped from connected networks since it
represents a successful TCP configuration on the host.

Signed-off-by: acikgozb <acikgozb@proton.me>