feat(toggle): Implement `wl toggle`
`wl toggle` (or `wl t` with alias) is implemented by getting the current
wifi status, and calling `nmcli::toggle_wifi` to change it.

Upon a successful execution, the latest status is printed to stdout.
Upon a failed execution, the `io::Error` is delegated to the caller.

The error messages and appropriate exit codes will be done in a separate
commit.

Signed-off-by: acikgozb <acikgozb@proton.me>