refactor(nmcli): Add support for multiple net be's
Even though `nmcli` is used exclusively in `wl`, it takes not much to
add support for multiple network backends.

Therefore, a new module is introduced in lib crate - `adapter` - which
holds the `Wl` trait that should be implemented by each network backend.

With this trait introduced, the flow has changed in different ways:

- Instead of initializing `Nmcli` in its own module, it is initialized
under the lib crate with the new `new()` method in each individual
subcommand.

- `Nmcli` struct and the functions are simplified and functions are
moved under the trait implementation.

Signed-off-by: acikgozb <acikgozb@proton.me>