feat(disconnect): Implement `wl disconnect`
`wl disconnect` (or `wl d` with alias) is designed to cover multiple
network disconnections and removals.

Since entering SSID by hand is not a welcoming experience, `wl
disconnect` is implemented with a small interactivity. If users do not
provide an SSID, it shows the active connections to let users choose.

As decided before, `--forget` flag is implemented to remove the selected
network from the known network list. Whilst keeping a network in known
network list makes it a lot easier and faster to connect to it, a hard
delete is required in some cases (e.g. when the password of the network
changes).

Signed-off-by: acikgozb <acikgozb@proton.me>