refactor(nmcli): Move ls output printing to nmcli
The callers of `nmcli` module should not know the format of the returned String's.

In `wl::list_networks()`, the process result `Vec<String>` is joined with newlines. This essentially means that the caller of `process.list_networks()` know that the vector actually represents lines of networks.

However, this can't be figured out from the signature of `process.list_networks()` alone.

Therefore, the responsibility of listing the networks is moved back to `nmcli` module. Lib crate should stay lean, and the format of things that are printed to stdout should be decided at network backend modules.

Signed-off-by: acikgozb <acikgozb@proton.me>