refactor(encoding): Support non UTF8 inputs
Up to this point, both the `wl` args and the output of the underlying network backend (`nmcli`) were assumed to be encoded in UTF-8.

However, this is not necessarily true all the time simply because SSID's may contain non UTF-8 values. This effects the whole project since SSID's are used pretty much in every `wl` subcommand.

Right now, if a non UTF-8 SSID is retrieved from `nmcli`, the corresponding `wl` subcommand fails because `wl` expects valid UTF-8 output from `nmcli`.

Therefore, the whole project is updated to support both UTF-8 and non UTF-8 values by processing `u8` instead of `String`'s.

However, this change does not support Windows platforms, but that is an acceptable tradeoff since `wl` is designed to work on Unix hosts only.

Signed-off-by: acikgozb <acikgozb@proton.me>