refactor(nmcli): Use tuple to for ssid/dev pairs
The first version of `nmcli` returned `Vec<String>`, which is joined
by lib crate (caller) via ", ".

While this resulted in a correct ssid-dev pair format, using ", "
indicated that the caller knows the actual format of `Vec<Vec<String>>`
which cannot be decided from the `Wl::get_active_ssid_dev_pairs`
signature only.

To resolve this ambiguity and mix of responsibilities about the return
type/format, a new type is introduced in `adapter` - `SsidDevPair` -
which is a tuple with 2 elements.

Signed-off-by: acikgozb <acikgozb@proton.me>