refactor(wl): Create submodules for lib and bin
Since `wl scan` takes a complex type as an argument, it was not possible to bring that type into lib crate's scope since the type was defined in the binary crate.

Therefore, a new module `api` is defined under the library crate to contain the `clap` structs and reference them freely in subcommands if needed.

While we are at it, the library crate is divided into submodules based on each subcommand as well.

Signed-off-by: acikgozb <acikgozb@proton.me>