feat(adapter): Add struct Decimal for signal cmp
Since `scan` will filter out the available networks based on their
signal strengths, a new wrapper type `Decimal` is added to the `adapter`
module.

This type wraps `&[u8]` to convert it to decimal `u8`, which helps network backends to compare the signal strengths coming from both the caller and the backend itself.

Signed-off-by: acikgozb <acikgozb@proton.me>