feat(lib): Update write_out to take impl io::Write
During the development of `wl scan`, it is realized that not all
subcommands will print their outputs to stdout.

For example, `wl connect` will use `wl::scan` in the absence of an SSID,
and in that case the scan list will be updated with indexes.

So the contents of the list will be written to an in memory list for `wl
connect`, not `stdout`.

Therefore, `crate::write_out` is implemented to take a type that
implements `io::Write`.

All usages are updated accordingly with `io::stdout` for now.

Signed-off-by: acikgozb <acikgozb@proton.me>