feat(scan): Implement `wl scan`
The subcommand `wl scan` (or `wl sc` with alias) is designed to scan
available WiFi networks on the host.

In addition to the functionality provided by `nmcli`, a filtering is
implemented based on each network's signal strength.
Since `nmcli` does not provide filtering, it is done by taking advantage
of the caching done on the list by `nmcli`.

However, the filtering is unfortunately non-deterministic since the
cache may expire during the process which may result in user query
showing different results than the filter query.

To decrease the possibilities of this happening, it is advised to use
signal filtering with the `--re-scan` flag to refresh the cache before
the filtering operation.

Signed-off-by: acikgozb <acikgozb@proton.me>