feat(api): Remove `ScanArgs` from `wl connect`
Having `ScanArgs` in `wl connect` allows users to filter and get values that may not contain the SSID.
So, by nature, user supplied arguments cannot be trusted in `wl connect` flow to show the SSID's.
This leads to querying the network backend a second time to ensure that the SSID's are retrieved and shown to the user.

The default network backend, `nmcli` gives back non-deterministic query results if it is queried for multiple scans.

Therefore, to have a more deterministic flow, `ScanArgs` is removed from `wl connect`.

`wl connect` will show its own scan list to the user to ensure that SSID's are always shown and retrieved (if the `-i` flag is not provided).

Signed-off-by: acikgozb <acikgozb@proton.me>