feat(adapter): Add get_field_separator to trait Wl
Network backends can be utilized to get a terse output, which can be
used for scripting.
However, terse output is constructed by using a field separator, and
this separator may change based on each network backend.

Since terse outputs are mainly used by the caller, it is impossible to
know how they will be used from the backend side.

Therefore, it is not beneficial to handle the actual parsing logic
inside the network backend.
For needs like these, the network backend should provide the field
separator that it uses to create the terse output. With this at hand,
the caller can manipulate the output however they prefer.

Therefore, a new method is added to `adapter::Wl` to obtain the field
separator.

Signed-off-by: acikgozb <acikgozb@proton.me>