feat(connect): Implement `wl connect`
`wl connect` (or `wl c` with alias) is designed to connect to both an existing network or a new one.

It is also designed to handle the unintuitive way of entering the SSID's by hand by providing an interactive mode.
If the SSID is not given by the user, then `wl c` shows a list of available networks with their signal strengths to choose from.

It is also designed refresh the network connection for the cases when the password of the network changes by using `--force-passwd`.

Signed-off-by: acikgozb <acikgozb@proton.me>